requirements.txt
requests==2.31.0
google-api-python-client==2.150.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
oauth2client==4.1.3
pandas==2.2.2
